Create export_figure.m
export_fig may cause compatibility issues as it requires the installation of some external software (ghostscript, Xpdf) some of which is mainly for linux. In order to avoid them it is better to let the user decide if he/she wants to use this function or the MATLAB's native ones.